fix inference datatype and generation config

+++ src/peft/tuners/lora.py
@@ -772,7 +772,7 @@ if is_bnb_available():
                 result = result.clone()
                 if not torch.is_autocast_enabled():
                     expected_dtype = result.dtype
-
+                    x = x.to(self.lora_A[self.active_adapter].weight.dtype)
                     output = (
                         self.lora_B[self.active_adapter](
                             self.lora_A[self.active_adapter](self.lora_dropout[self.active_adapter](x))
